Evaluate number of connections and 'errors' string in health check
Previously a health check failed if and when zcash-cli exited with an
error code. For example, because zcashd did not respond.

This one is finer, and reports the container as unhealthy if for whatever
reason zcashd cannot establish enough connections, or reports any other
error message.

Warnings are skipped because a warning per definition is symptom of a
state in which a program can still work.